+++ src/main/java/com/advocate/dto/request/LoginRequest.java
@@ -1,0 +1,16 @@
+package com.advocate.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@ToString
+public class LoginRequest {
+    private String email;
+    private String password;
+    private String role;
+}
