+++ src/main/java/com/advocate/dto/request/SignupRequest.java
@@ -1,4 +1,4 @@
-package com.advocate.dto;
+package com.advocate.dto.request;
 
 import com.advocate.entity.Address;
 import com.fasterxml.jackson.annotation.JsonInclude;
