+++ src/main/java/com/advocate/dto/request/UpdateClientRequestDto.java
@@ -1,4 +1,4 @@
-package com.advocate.dto;
+package com.advocate.dto.request;
 
 
 import com.advocate.entity.Address;
