+++ src/main/java/com/advocate/dto/response/CommonResponseDto.java
@@ -1,4 +1,4 @@
-package com.advocate.dto;
+package com.advocate.dto.response;
 
 import org.springframework.http.HttpStatus;
